requirements.txt
flask
flask-cors
python-dotenv
langchain==0.1.16
langchain-google-genai
google-api-python-client
google-auth
requests